Updated identity commitment generation class
